log message to tie together old and new name for tests
Since 03a75ecb32d0ef1fd45f2549b8cfeab77a05c693 we've been using test
names that are the same as the Clojure test defined in code.

The old test names are confusing, since it's hard to locate the clojure
test in the code that corresponds to the name.

Still, some old github issues refer to the old names, so it's good to
have a place to connect the two.

Signed-off-by: Yoav Kleinberger <yoav@scylladb.com>